fix: Loading the last payment after saving and continue entering
When checking the continue entering flg, after save the current
transaction, the first payment was being loaded and the field was not
being updated. The better fix is load the last payment selected, in most
of cases the person wants to keep the the same payment and category, so
is not needed to reset to the fisrt value.

Issue: https://github.com/GeovaniTech/InvestMe/issues/63